Moved extractIperfResults to utils.js; Checked that it functions and passes tests on macOS
